+++ job-post-delete.php
@@ -1,4 +1,9 @@
 <?php
+session_start();
+if (!isset($_SESSION['admin_name']) && !isset($_SESSION['super_admin_name'])) {
+    header('Location: index.php');
+    exit();
+}
 include "db_conn.php";
 $id = $_GET['id'];
 $sql = "DELETE FROM `bcp-sms3_job` WHERE id = $id";
@@ -11,3 +16,4 @@ else{
     echo "Failed:" . mysqli_error($conn);
 }
 ?>
+
